Cleaning, Fixed a bug in the class generator (Preselection)
